+++ SAP/src/SAP/obj/Debug/net8.0-windows/win-x64/SAP.AssemblyInfo.cs
@@ -15,7 +15,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCopyrightAttribute("Copyright © 2024-2025")]
 [assembly: System.Reflection.AssemblyDescriptionAttribute("Allocation Buddy, Essentials Buddy, and ExpireWise inventory management tools")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("4.1.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("4.1.0+3fb51cbace8ee4c1c0c91fd8b6214ac52e9037ac")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("4.1.0+a6cba827821e848bca8c2de8600b05afd8e82fb6")]
 [assembly: System.Reflection.AssemblyProductAttribute("S.A.P - S.A.M. Add-on Pack")]
 [assembly: System.Reflection.AssemblyTitleAttribute("SAP")]
 [assembly: System.Reflection.AssemblyVersionAttribute("4.1.0.0")]